[#36] RNNCell are now optimised with parallelism
