+++ cleanrnns/pipelines.py
@@ -1,7 +1,7 @@
 import torch  # noqa
 from typing import Tuple, List
 from tokenizers import Tokenizer
-from cleanrnns.models import ClassificationBase
+from cleanrnns.modules import ClassificationBase
 from cleanrnns import tensors as T  # noqa
