+++ explore/explore_cleanrnns_fetch_model_for_classification.py
@@ -1,9 +1,9 @@
 import shutil
-from fetchers import fetch_model_for_classification
+from fetchers import fetch_module_for_classification
 
 
 def main():
-    model = fetch_model_for_classification("rnn_for_classification")
+    model = fetch_module_for_classification("rnn_for_classification")
     print(model.hparams)  # just a sanity check
     shutil.rmtree("artifacts")  # clear the cache after testing
